+++ source/entities/bullet.h
@@ -7,7 +7,7 @@
 typedef struct {
     int x, y;          // Posição (Ponto Fixo)
     int dx, dy;        // Vetores de Velocidade (Ponto Fixo)
-    int angle;         // Índice da LUT (0-31)
+    u8 angle;         // Índice da LUT (0-31)
     OBJATTR *obj;      // Ponteiro para o atributo na OAM
     bool active;      // Flag para controle de existência (útil para tiros e asteroides)
 } Bullet;
